+++ frontend/projects/api/src/lib/deployment/version.service.token.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+import { IVersionService } from './version.service.interface';
+
+export const VERSION_SERVICE = new InjectionToken<IVersionService>('VERSION_SERVICE');